# mrcomic-api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mr.Comic API",
    "description": "API для сторонних разработчиков платформы Mr.Comic",
    "version": "1.0.0",
    "contact": {
      "name": "Mr.Comic API Support",
      "email": "api-support@mrcomic.com",
      "url": "https://developer.mrcomic.com"
    }
  },
  "servers": [
    {
      "url": "https://api.mrcomic.com/v1",
      "description": "Продакшн сервер"
    },
    {
      "url": "https://api-sandbox.mrcomic.com/v1",
      "description": "Песочница для тестирования"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/themes": {
      "get": {
        "summary": "Получение списка тем",
        "description": "Возвращает список пользовательских тем с пагинацией и возможностью фильтрации",
        "operationId": "getThemes",
        "tags": ["Themes"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Номер страницы",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Количество элементов на странице",
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Поисковый запрос",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Поле для сортировки",
            "schema": {
              "type": "string",
              "enum": ["name", "createdAt", "rating", "downloads"],
              "default": "rating"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Порядок сортировки",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешный ответ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThemesList"
                }
              }
            }
          },
          "400": {
            "description": "Неверный запрос",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/themes/{id}": {
      "get": {
        "summary": "Получение информации о теме",
        "description": "Возвращает детальную информацию о теме по её ID",
        "operationId": "getThemeById",
        "tags": ["Themes"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID темы",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешный ответ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Theme"
                }
              }
            }
          },
          "404": {
            "description": "Тема не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Theme": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "isPublic": {
            "type": "boolean"
          },
          "previewImageUrl": {
            "type": "string",
            "format": "uri"
          },
          "authorId": {
            "type": "string"
          },
          "author": {
            "$ref": "#/components/schemas/User"
          },
          "rating": {
            "type": "object",
            "properties": {
              "average": {
                "type": "number",
                "format": "float"
              },
              "count": {
                "type": "integer"
              }
            }
          },
          "downloads": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "themeData": {
            "type": "object",
            "properties": {
              "colors": {
                "type": "object",
                "properties": {
                  "primary": {
                    "type": "string"
                  },
                  "secondary": {
                    "type": "string"
                  },
                  "accent": {
                    "type": "string"
                  },
                  "background": {
                    "type": "string"
                  },
                  "text": {
                    "type": "string"
                  }
                }
              },
              "fonts": {
                "type": "object",
                "properties": {
                  "main": {
                    "type": "string"
                  },
                  "headers": {
                    "type": "string"
                  },
                  "size": {
                    "type": "string"
                  }
                }
              },
              "elements": {
                "type": "object",
                "properties": {
                  "cornerRadius": {
                    "type": "integer"
                  },
                  "iconStyle": {
                    "type": "string"
                  },
                  "buttonStyle": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["id", "name", "authorId", "themeData"]
      },
      "ThemesList": {
        "type": "object",
        "properties": {
          "themes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Theme"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "required": ["themes", "pagination"]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "avatarUrl": {
            "type": "string",
            "format": "uri"
          },
          "bio": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": ["user", "moderator", "admin"]
          },
          "reputation": {
            "type": "integer"
          },
          "preferences": {
            "type": "object",
            "properties": {
              "notifications": {
                "type": "boolean"
              },
              "language": {
                "type": "string"
              }
            }
          }
        },
        "required": ["id", "username", "role"]
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "pages": {
            "type": "integer"
          }
        },
        "required": ["total", "page", "limit", "pages"]
      },
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object"
          }
        },
        "required": ["code", "message"]
      }
    }
  }
}